requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2022.1
Django==4.2.20
django-ckeditor==6.7.2
django-ckeditor-5==0.2.17
django-cors-headers==4.7.0
django-crispy-forms==1.14.0
django-debug-toolbar==5.1.0
django-js-asset==2.2.0
django-ratelimit==4.1.0
pillow==11.2.1
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.4.0
whitenoise==6.6.0